web-manager/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
docker==6.1.3
pyyaml==6.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
requests==2.31.0
psutil==5.9.6
